tests/submission/Microsoft/video/video_027.htm

<!DOCTYPE HTML>
<html>
<head>
  <title>HTML5 Media Elements: The inherited 'seeking' property of the 'HTMLVideoElement' interface is read-only</title>
  <meta content="text/html; charset=UTF-8" http-equiv="Content-Type">
  <link rel="author" title="Microsoft" href="http://www.microsoft.com/" />
  <link rel="help" href="http://www.w3.org/TR/html5/video.html#dom-media-seeking" />
  <meta name="assert" content="The inherited 'seeking' property of the 'HTMLVideoElement' interface is read-only" />
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script> 
</head>
<body>
  <div id=log></div>
  <video id="video0">Your browser does not support media elements.</video>
  <script type="text/javascript">

    var testElem = document.getElementById("video0");
    testElem.seeking = true;
  
    test(function() { assert_false(testElem.seeking, "The inherited 'seeking' property of the 'HTMLVideoElement' interface is read-only")});
  
  </script>
</div>
</body>
</html>
